requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-nested-admin==4.1.1
django-rest-framework==0.1.0
#django-tapsi-auth-ldap==0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
executing==2.1.0
Faker==28.4.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
ldap3==2.9.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13
